+++ pom.xml
@@ -48,7 +48,6 @@
         <plexus-utils.version>3.5.1</plexus-utils.version>
 
         <!-- Test -->
-        <spek-api.version>1.2.1</spek-api.version>
         <mockito-kotlin.version>1.6.0</mockito-kotlin.version>
         <junit-platform-surefire-provider.version>1.3.2</junit-platform-surefire-provider.version>
 
@@ -113,16 +112,17 @@
         </dependency>
 
         <dependency>
-            <groupId>org.jetbrains.spek</groupId>
-            <artifactId>spek-api</artifactId>
-            <version>${spek-api.version}</version>
+            <groupId>org.jetbrains.kotlin</groupId>
+            <artifactId>kotlin-test</artifactId>
+            <version>${kotlin.version}</version>
             <scope>test</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.jetbrains.kotlin</groupId>
-                    <artifactId>kotlin-reflect</artifactId>
-                </exclusion>
-            </exclusions>
+        </dependency>
+
+        <dependency>
+            <groupId>org.jetbrains.kotlin</groupId>
+            <artifactId>kotlin-test-junit5</artifactId>
+            <version>${kotlin.version}</version>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>com.beust</groupId>
@@ -130,18 +130,6 @@
             <version>${jcommander.version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.jetbrains.spek</groupId>
-            <artifactId>spek-junit-platform-engine</artifactId>
-            <version>${spek-api.version}</version>
-            <scope>test</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.jetbrains.kotlin</groupId>
-                    <artifactId>kotlin-reflect</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
             <artifactId>kotlin-reflect</artifactId>
@@ -278,7 +266,7 @@
                 <version>${maven-surefire-plugin.version}</version>
                 <configuration>
                     <includes>
-                        <include>**/*Spec.*</include>
+                        <include>**/*Test.*</include>
                     </includes>
                 </configuration>
                 <dependencies>
@@ -288,15 +276,9 @@
                         <version>${junit-platform-surefire-provider.version}</version>
                     </dependency>
                     <dependency>
-                        <groupId>org.jetbrains.spek</groupId>
-                        <artifactId>spek-junit-platform-engine</artifactId>
-                        <version>${spek-api.version}</version>
-                        <exclusions>
-                            <exclusion>
-                                <groupId>org.jetbrains.kotlin</groupId>
-                                <artifactId>kotlin-runtime</artifactId>
-                            </exclusion>
-                        </exclusions>
+                        <groupId>org.jetbrains.kotlin</groupId>
+                        <artifactId>kotlin-test-junit5</artifactId>
+                        <version>${kotlin.version}</version>
                     </dependency>
                 </dependencies>
             </plugin>
@@ -375,6 +357,11 @@
             <name>JCenter Repository</name>
             <url>https://jcenter.bintray.com/</url>
         </repository>
+        <repository>
+            <id>maven</id>
+            <name>Maven</name>
+            <url>https://repo1.maven.org/maven2/</url>
+        </repository>
     </repositories>
 
     <pluginRepositories>
@@ -382,6 +369,11 @@
             <id>jcenter</id>
             <name>JCenter Repository</name>
             <url>https://jcenter.bintray.com/</url>
+        </pluginRepository>
+        <pluginRepository>
+            <id>maven</id>
+            <name>Maven</name>
+            <url>https://repo1.maven.org/maven2/</url>
         </pluginRepository>
     </pluginRepositories>
